breakdown crud into multiple modules
